The first stop on a long journey
Add the beginnings of a role to template some files to be used to
build a docker image that will run ansible as a build step to configure
the image.

A seperate role will be used to perform the build step, to give consumers
a chance to add any additional configuration to the image build.